feat: 14 remove google books api parser class
