+++ google_books_api_wrapper/api.py
@@ -1,7 +1,7 @@
 import logging
 from .rest_adapter import RestAdapter
 from .constants import GOOGLE_BOOKS_API_URL
-from .models import Book, HttpResult, BookSearchResultSet, GoogleBooksSearchParams
+from .models import Book, BookSearchResultSet, GoogleBooksSearchParams
 
 class GoogleBooksAPI:
     """Wrapper around the Google Books REST API
@@ -60,7 +60,7 @@ class GoogleBooksAPI:
                 subject=subject,
             ).generate(),
         )
-        result_set = GoogleBooksApiParser.get_books_from_response(response)
+        result_set = BookSearchResultSet.from_google_books_api_response(response.data)
         return result_set
 
     def get_book_by_isbn13(self, isbn13: int) -> Book:
@@ -105,7 +105,7 @@ class GoogleBooksAPI:
             ep_params=GoogleBooksSearchParams(
                 subject=subject
             ).generate())
-        result_set = GoogleBooksApiParser.get_books_from_response(response)
+        result_set = BookSearchResultSet.from_google_books_api_response(response.data)
         return result_set
     
     def _get_book_by_isbn(self, isbn_num) -> Book:
@@ -114,28 +114,5 @@ class GoogleBooksAPI:
         endpoint="volumes",
         ep_params=GoogleBooksSearchParams(isbn=isbn_num).generate(),
         )
-        result_set = GoogleBooksApiParser.get_books_from_response(response)
+        result_set = BookSearchResultSet.from_google_books_api_response(response.data)
         return result_set.get_best_match()
-
-
-class GoogleBooksApiParser:
-    @staticmethod
-    def get_books_from_response(response: HttpResult) -> BookSearchResultSet:
-        book_results_from_web_api = (
-            response.data["items"] if "items" in response.data else []
-        )
-        book_results = [
-            Book.from_api_response_item(book_result)
-            for book_result in book_results_from_web_api
-        ]
-        return BookSearchResultSet(books=book_results)
-    
-
-    @staticmethod
-    def get_isbn_from_id_list(
-        industry_ids: list[dict[str, str]], *, isbn_num: int
-    ) -> str:
-        for id in industry_ids:
-            if id["type"] == "ISBN_" + str(isbn_num):
-                return id["identifier"]
-        return None
