+++ tests/unit/test_rest_adapter.py
@@ -3,7 +3,6 @@ import responses
 from google_books_api_wrapper.rest_adapter import RestAdapter
 from google_books_api_wrapper.constants import GOOGLE_BOOKS_API_URL
 from google_books_api_wrapper.exceptions import GoogleBooksAPIException
-import logging
 import pytest
